Early addition / Work
Added in pro-builder to allow for the cones that you see me using for the boids
I am also now using a combination of both Sebastion Lague & another tutorial that I saw online made by "Dawn Studio."
I can already tell however, that Dawn Studio's methods are for sure going to cause performance issues. But that'll be an issue I address later, likely through use of putting most of the load on the GPU through computing.