The 3 Fundamentals Implemented
I've added in:
Separation, Alignment, & Cohesion!

Currently, this isn't the most optimal approach as I can already see that the CPU is getting bottlenecked trying to compute all of the For loops for each individual Boid.
This will be something that I fix later through the use of GPU threading, but before I get to that I need to finish one last thing.
That being, of course, obstacle avoidance. Right now we're forcing the boids to remain in their container through teleporting them back when they escape out of bounds. However we wish to have them not even attempt to leave at all by simply being aware that there's a wall in front of them. I already have an idea of how to do this so it'll be the next thing that i'll do.