Technically done, needs reworks/optimizations
While this technically gets across the idea of what we want the boids to do, it could use a lot of improvements.
One thing that for sure needs rework is the obstacle detection system, as simply casting a single raycast forwards doesn't create a natural avoidance system. The boids also fail to avoid their obstacles fast enough due to their rotation speed / regular speed not being reactive. This is something that Sebastion Lague delt with, and I think this is the point where I need to look at his approach for things.
I am aware that this will involve a lot of math, and will for sure cause a lot of computational overhead. But we can worry about the performance once the boids are functional.